add VTX_MEM_CACHE_INVALIDATE and VTX_GPU_CACHE_INVALIDATE bits
After fiddling a bit more with the DRAW_PARAMS bits, I'm now more convinced
that there are two caches for vertices:

- one is external memory cache, containing pre-fetched data
- second is internal GPU cache, containing pre-processed(?) data

It's easy to see that stale vertex data is used if one of the bits is unset.

Seems the GPU cache contains a parsed vertex data that is binned
per triangle's vertex, and thus, that data needs to be invalidated
if new drawing operation uses different number of attributes per-vertex,
but the data in memory cache is not changed since the time of a previous
drawing operation.

For example I see that the VTX_GPU_CACHE_INVALIDATE needs to be set by
Opentegra driver when number of attributes used per-vertex is changed on
a second draw call, but vertex data itself is unchanged since the first
draw call. While VTX_MEM_CACHE_INVALIDATE needs to be set if vertex data
in the memory buffer is changed.